finished the post route for a new offer

+++ backend/db/migrate/20150306203359_create_listings.rb
@@ -1,8 +1,8 @@
 class CreateListings < ActiveRecord::Migration
   def change
     create_table :listings do |t|
-      t.integer :max_price
-      t.integer :accept_price
+      t.float :max_price
+      t.float :accept_price
       t.belongs_to :user
       t.belongs_to :product
       t.datetime :expires_at
