+++ backend/db/migrate/20150306203513_create_offers.rb
@@ -1,7 +1,7 @@
 class CreateOffers < ActiveRecord::Migration
   def change
     create_table :offers do |t|
-      t.integer :offer_price
+      t.float :offer_price
       t.belongs_to :user
       t.belongs_to :product
       t.integer :bid_id
